* 2018.02.26 오후 16:44 수정할 사항
1. Class name Refactoring
2. Change SensorData.dbml, associated relation to extends relation
3. Refactoring extra coding caused by number2

+++ WindowForm_Dongik/SensorAddForm.cs
@@ -144,7 +144,7 @@ namespace WindowForm_Dongik
             {
                 Ip = this.IpTextBox.Text,
                 Port = Convert.ToInt32(this.PortTextBox.Text),
-                Address = Convert.ToInt32(this.AddressTextBox.Text)
+                Address = Convert.ToUInt16(this.AddressTextBox.Text)
             };
             dbManager.dc.ModbusSensors.InsertOnSubmit(newConfig.ModbusTables);
         }
